NoteVault-main/notevaultBackend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djongo
dnspython==2.6.1
fonttools==4.49.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
openai==1.51.0
packaging==23.2
pillow==10.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==3.12.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
sockets==1.0.0
sqlparse==0.2.4
threadpoolctl==3.3.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
